Moved non-Phyloref stuff back into App.vue for better code organization.
